// src/app/modules/order-return/models/LrMismatchCountModel.ts
export class LrMismatchCountModel {
  BranchId: number = 0;
  CompId: number = 0;
  TodayTotalLR: number = 0;
  TodayReceivedLR: number = 0;
  TodayImportedLR: number = 0;
  NotFoundLR: number = 0;
  LREntryId: number = 0;
  TodayLRGP: number = 0;
  ConcernCnt: number = 0;
  ConcernResolveCnt: number = 0;
  RecvdAtOPCnt: number = 0;
  PendingAtExpSCnt: number = 0;
  TodayVerifiedCount: number = 0;
  CorrReqCount: number = 0;
}